Refactor collections: replace CollectionPage with CollectionDetailPage and update routing
